pyproject.toml
[build-system]
requires = ["setuptools>=77", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "genesis-sensors"
version = "0.1.0"
description = "Self-contained sensor rigs and demos for Genesis"
readme = "README.md"
requires-python = ">=3.10"
license = "Apache-2.0"
license-files = ["LICENSE"]
authors = [{name = "Vincent Lordier", email = "vlordier@users.noreply.github.com"}]
keywords = ["genesis", "robotics", "sensors", "simulation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
dependencies = [
    "genesis-world>=0.4.2",
    "numpy>=1.26.4",
    "pydantic>=2.11.0",
]

[project.optional-dependencies]
dev = [
    "build>=1.2.2",
    "bump-my-version>=0.32.1",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.5.39",
    "mkdocstrings[python]>=0.29",
    "poethepoet>=0.32.0",
    "pre-commit>=4.0.1",
    "pytest>=8.0",
    "rerun-sdk>=0.23",
    "python-semantic-release>=10.2.0",
    "ruff>=0.9",
    "twine>=6.1.0",
]
rerun = [
    "rerun-sdk>=0.23",
]

[project.urls]
Homepage = "https://vlordier.github.io/genesis-sensors/"
Repository = "https://github.com/vlordier/genesis-sensors"
Issues = "https://github.com/vlordier/genesis-sensors/issues"
Documentation = "https://vlordier.github.io/genesis-sensors/"
Changelog = "https://github.com/vlordier/genesis-sensors/releases"

[project.scripts]
genesis-sensors = "genesis_sensors.cli:main"
gs-sensors = "genesis_sensors.cli:main"
genesis-sensors-demo = "genesis_sensors.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["genesis_sensors", "genesis_sensors.*"]

[tool.setuptools.package-data]
genesis_sensors = ["py.typed"]

[tool.ruff]
line-length = 120

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = ["-q"]

[tool.bumpversion]
current_version = "0.1.0"
parse = "(?P<major>\\d+)\\.(?P<minor>\\d+)\\.(?P<patch>\\d+)"
serialize = ["{major}.{minor}.{patch}"]
commit = false
tag = false
message = "chore(release): bump version {current_version} -> {new_version}"

[[tool.bumpversion.files]]
filename = "pyproject.toml"

[[tool.bumpversion.files]]
filename = "src/genesis_sensors/__init__.py"
search = "__version__ = \"{current_version}\""
replace = "__version__ = \"{new_version}\""

[tool.semantic_release]
allow_zero_version = true
build_command = "python -m pip install --upgrade build && python -m build"
commit_message = "chore(release): {version} [skip ci]"
commit_parser = "conventional"
exclude_commit_patterns = ["^chore\\(release\\):"]
major_on_zero = false
tag_format = "v{version}"
upload_to_vcs_release = true
version_toml = ["pyproject.toml:project.version"]
version_variables = ["src/genesis_sensors/__init__.py:__version__"]

[tool.semantic_release.commit_parser_options]
allowed_tags = ["feat", "fix", "perf", "build", "ci", "docs", "refactor", "test", "chore"]
minor_tags = ["feat"]
patch_tags = ["fix", "perf"]
parse_squash_commits = true
ignore_merge_commits = true

[tool.semantic_release.branches.main]
match = "main"
prerelease = false

[tool.semantic_release.changelog.default_templates]
changelog_file = "CHANGELOG.md"
output_format = "md"

[tool.poe.tasks]
test = { shell = "PYTHONPATH=src pytest tests -q" }
fault_demo = { shell = "PYTHONPATH=src python examples/fault_tolerant_sensors.py --frames 8 --dropout 0.2" }
environment_demo = { shell = "PYTHONPATH=src python examples/environmental_sensors.py --frames 8" }
wireless_demo = { shell = "PYTHONPATH=src python examples/sensor_usage_patterns.py --frames 8" }
doc_assets = { shell = "PYTHONPATH=src python examples/generate_sensor_doc_assets.py --output-dir docs/assets/sensors" }
rerun_demo = { shell = "PYTHONPATH=src python examples/external_sensors_rerun.py --frames 60 --spawn" }
rerun_save = { shell = "PYTHONPATH=src python examples/external_sensors_rerun.py --frames 60 --save /tmp/genesis-sensors-traces.rrd" }
check = { shell = "pre-commit run --all-files && PYTHONPATH=src pytest tests -q" }
